Seperate the DB class from the Consumer class * create class to support DB for consumer * update FoxConsumer class
