+++ fox_consumer/__init__.py
@@ -1,2 +1,3 @@
 from fox_helper import FoxItem
 from consumer import FoxConsumer
+from consumer_db_support import FoxConsumerPostgres
